refactor: Update image URLs and remove Koin modules in mediaDetailsUi
This commit updates the image URL construction in `TvShowsMapper.kt`, `TvShowLogoDto.kt`, `TvShowCastDto.kt`, and `TvShowDto.kt` to use a consistent base URL.

Additionally, it removes the Koin `RepositoryModule` and `UseCaseModule` from the `mediaDetailsUi` feature module.

Other changes include:
- Removing alpha from fill colors in several drawable icons.
- Updating `NetworkModule` to provide multiple `NetworkConnectionChecker` instances for different repository modules.
- Adjusting `DescriptionSection` to conditionally render genres and description.
- Ensuring `RatingCard` in `DetailsImage` receives a Float for the rating.
- Adding logging in `MediaDetailsViewModel` and `MovieRepositoryImpl` for debugging purposes.
- Instantiating multiple `NetworkConnectionChecker` instances in `AflamiApplication`.

+++ feature/mediaDetails/mediaDetailsUi/src/main/java/com/feature/mediaDetails/mediaDetailsUi/ui/comon/components/detailsImage/DetailsImage.kt
@@ -97,7 +97,7 @@ fun DetailsImage(
                 }
             }
             RatingCard(
-                rating = rating.toFloat(),
+                rating = rating,
                 modifier = Modifier
                     .align(Alignment.BottomStart)
                     .padding(start = 4.dp)
